use mantine select, better unique column identifiers
