+++ frontend/src/metabase/common/components/FilterPicker/CoordinateFilterPicker/constants.ts
@@ -1,4 +1,3 @@
-import { t } from "ttag";
 import type { CoordinateFilterValueCountMap } from "./types";
 
 // why doesn't mbql recognize is-null not-null for coordinates?
@@ -12,10 +11,3 @@ export const coordinateFilterValueCountMap: CoordinateFilterValueCountMap = {
   "=": "multiple",
   "!=": "multiple",
 };
-
-export const insideLabels = [
-  t`Upper latitude`,
-  t`Left longitude`,
-  t`Right longitude`,
-  t`Lower latitude`,
-];
